Data.BCP47

Description

Human beings on our planet have, past and present, used a number of languages. There are many reasons why one would want to identify the language used when presenting or requesting information.

The language of an information item or a user's language preferences often need to be identified so that appropriate processing can be applied. For example, the user's language preferences in a Web browser can be used to select Web pages appropriately. Language information can also be used to select among tools (such as dictionaries) to assist in the processing or understanding of content in different languages. Knowledge about the particular language used by some piece of information content might be useful or even required by some types of processing, for example, spell-checking, computer-synthesized speech, Braille transcription, or high-quality print renderings.

- https://tools.ietf.org/html/bcp47

Documentation

data BCP47

A language tag

Language tags are used to help identify languages, whether spoken, written, signed, or otherwise signaled, for the purpose of communication. This includes constructed and artificial languages but excludes languages not intended primarily for human communication, such as programming languages.

inits :: BCP47 -> [BCP47]

Produce a list of (<= priority) language tags

>>> inits enGBTJP
[en,en-GB,en-GB-t-jp]

Construction

mkLanguage :: ISO639_1 -> BCP47

Construct a simple language tag

mkLocalized :: ISO639_1 -> Country -> BCP47

Construct a localized tag

fromText :: Text -> Either Text BCP47

Parse a language tag from text

>>> fromText $ pack "en"
Right en

>>> fromText $ pack "de-CH"
Right de-CH

>>> fromText $ pack "ru-USR"
Left "fromText:1:3:\n  |\n1 | ru-USR\n  |   ^\nunexpected '-'\n"

>>> fromText $ pack "en-a-ccc-v-qqq-a-bbb"
Right en-a-bbb-a-ccc-v-qqq

>>> fromText $ pack "de-Latn-DE"
Right de-Latn-DE

>>> fromText $ pack "de-Latf-DE"
Right de-Latf-DE

>>> fromText $ pack "de-CH-1996"
Right de-CH-1996

>>> fromText $ pack "de-Deva"
Right de-Deva

>>> fromText $ pack "zh-Hant-CN-x-private1-private2"
Right zh-Hant-CN-x-private1-private2

>>> fromText $ pack "zh-Hant-CN-x-private1"
Right zh-Hant-CN-x-private1

>>> fromText $ pack "zh-Hant-CN"
Right zh-Hant-CN

>>> fromText $ pack "zh-Hant"
Right zh-Hant

>>> fromText $ pack "zh"
Right zh

Serialization

toText :: BCP47 -> Text

Serialize BCP47 to Text

Subtags are serialized in the order described in the BCP 47 specification. Private-use subtags only appear at the end prefixed with an x.

Subtags

A language tag is composed from a sequence of one or more "subtags", each of which refines or narrows the range of language identified by the overall tag. Subtags, in turn, are a sequence of alphanumeric characters (letters and digits), distinguished and separated from other subtags in a tag by a hyphen ("-", [Unicode] U+002D).

toSubtags :: BCP47 -> [Subtags]

Convert tag to list of subtags

Language

data ISO639_1

language :: BCP47 -> ISO639_1

The language subtag

languageToText :: ISO639_1 -> Text

languageFromText :: Text -> Either Text ISO639_1

Parse a language subtag from Text

Language Extension

data LanguageExtension

Extended language subtags

These are used to identify certain specially selected languages that, for various historical and compatibility reasons, are closely identified with or tagged using an existing primary language subtag.

extendedLanguageSubtags :: BCP47 -> Set LanguageExtension

Look up all language extension subtags

languageExtensionToText :: LanguageExtension -> Text

languageExtensionFromText :: Text -> Either Text LanguageExtension

Parse a LanguageExtension subtag from Text

Language Script

data Script

Script subtags

Script subtags are used to indicate the script or writing system variations that distinguish the written forms of a language or its dialects.

script :: BCP47 -> Maybe Script

Look up the script subtag

scriptToText :: Script -> Text

scriptFromText :: Text -> Either Text Script

Parse a Script subtag from Text

Region

data Country

A country recognized by ISO 3166.

region :: BCP47 -> Maybe Country

Look up the region subtag

regionToText :: Country -> Text

regionFromText :: Text -> Either Text Country

Parse a region subtag from Text

>>> regionFromText $ pack "ZW"
Right zimbabwe

>>> regionFromText $ pack "012"
Right algeria

>>> regionFromText $ pack "asdf"
Left "regionFromText:1:1:\n  |\n1 | asdf\n  | ^\nunexpected 'a'\nexpecting 2 or 3 character country code\n"

Variant

data Variant

Variant subtags

Variant subtags are used to indicate additional, well-recognized variations that define a language or its dialects that are not covered by other available subtags.

variants :: BCP47 -> Set Variant

Look up all variant subtags

variantToText :: Variant -> Text

variantFromText :: Text -> Either Text Variant

Parse a Variant subtag from Text

Extension

data Extension

Extension subtags

Extensions provide a mechanism for extending language tags for use in various applications. They are intended to identify information that is commonly used in association with languages or language tags but that is not part of language identification.

extensions :: BCP47 -> Set Extension

Look up all extension subtags

extensionToText :: Extension -> Text

extensionFromText :: Text -> Either Text Extension

Parse an Extension subtag from Text

Private Use

data PrivateUse

Private Use subtags

Private use subtags are used to indicate distinctions in language that are important in a given context by private agreement.

privateUse :: BCP47 -> Set PrivateUse

Look up all private use subtags

privateUseToText :: PrivateUse -> Text

privateUseFromText :: Text -> Either Text (Set PrivateUse)

Parse a PrivateUse subtag from Text

For testing

en :: BCP47

English

es :: BCP47

Spanish

sw :: BCP47

Swahili

enGB :: BCP47

British English

enUS :: BCP47

American English

enTJP :: BCP47

A nonsense tag en-t-jp

enGBTJP :: BCP47

A nonsense tag en-GB-t-jp