Data.BCP47.Internal.Arbitrary

Documentation

class Arbitrary a

Random generation and shrinking of values.

QuickCheck provides Arbitrary instances for most types in base, except those which incur extra dependencies. For a wider range of Arbitrary instances see the quickcheck-instances package.

Minimal complete definition

arbitrary

arbitrary :: Arbitrary a => Gen a

A generator for values of the given type.

It is worth spending time thinking about what sort of test data you want - good generators are often the difference between finding bugs and not finding them. You can use sample, label and classify to check the quality of your test data.

There is no generic arbitrary implementation included because we don't know how to make a high-quality one. If you want one, consider using the testing-feat or generic-random packages.

The QuickCheck manual goes into detail on how to write good generators. Make sure to look at it, especially if your type is recursive!

alphaString :: Int -> Gen String

alphaNumString :: Int -> Gen String

alphaChar :: Gen Char

numChar :: Gen Char

elements :: [a] -> Gen a

Generates one of the given values. The input list must be non-empty.

choose :: Random a => (a, a) -> Gen a

Generates a random element in the given inclusive range.

oneof :: [Gen a] -> Gen a

Randomly uses one of the given generators. The input list must be non-empty.

suchThat :: Gen a -> (a -> Bool) -> Gen a

Generates a value that satisfies a predicate.

listOf :: Gen a -> Gen [a]

Generates a list of random length. The maximum length depends on the size parameter.

vectorOf :: Int -> Gen a -> Gen [a]

Generates a list of the given length.