Data.BCP47.Internal.Extension

Documentation

newtype Extension

Extension subtags

Extensions provide a mechanism for extending language tags for use in various applications. They are intended to identify information that is commonly used in association with languages or language tags but that is not part of language identification.

Constructors

Extension Text

extensionFromText :: Text -> Either Text Extension

Parse an Extension subtag from Text

extensionToText :: Extension -> Text

extensionP :: Parsec Void Text Extension

BCP-47 extension parser

@@ extension = singleton 1*("-" (2*8alphanum)) ; Single alphanumerics ; "x" reserved for private use

singleton = DIGIT ; 0 - 9 / %x41-57 ; A - W / %x59-5A ; Y - Z / %x61-77 ; a - w / %x79-7A ; y - z @@