Data.BCP47.Internal.Parser

Documentation

complete :: Parsec Void Text a -> Parsec Void Text a

Ensure a subtag extends to the next - or end of input

Used for subtags that can match some prefix of another subtag. For example, a Script or Region can accidentally be parsed from the prefix of a Variant

The alternative would be to use notFollowedBy with knowledge of the legal characters in the next valid subtag.