Data.BCP47.Internal.PrivateUse

Documentation

newtype PrivateUse

Private Use subtags

Private use subtags are used to indicate distinctions in language that are important in a given context by private agreement.

Constructors

PrivateUse Text

privateUseFromText :: Text -> Either Text (Set PrivateUse)

Parse a PrivateUse subtag from Text

privateUseToText :: PrivateUse -> Text

privateUseP :: Parsec Void Text (Set PrivateUse)

BCP-47 private use parser

@ privateuse = "x" 1*("-" (1*8alphanum)) @