Data.BCP47.Internal.Script

Documentation

newtype Script

Script subtags

Script subtags are used to indicate the script or writing system variations that distinguish the written forms of a language or its dialects.

Constructors

Script Text

scriptFromText :: Text -> Either Text Script

Parse a Script subtag from Text

scriptToText :: Script -> Text

scriptP :: Parsec Void Text Script

BCP-47 script parser

@ script = 4ALPHA ; ISO 15924 code @