Data.BCP47.Internal.Variant

Documentation

newtype Variant

Variant subtags

Variant subtags are used to indicate additional, well-recognized variations that define a language or its dialects that are not covered by other available subtags.

Constructors

Variant Text

variantFromText :: Text -> Either Text Variant

Parse a Variant subtag from Text

variantToText :: Variant -> Text

variantP :: Parsec Void Text Variant

BCP-47 variant parser

@ variant = 5*8alphanum ; registered variants / (DIGIT 3alphanum) @