Data.BCP47.Trie

Description

A trie like data structure for defining maps from BCP47 tags to values.

This structure supports collection and lookup of language tagged values. Its semantics are based on those defined in the BCP 47 specification.

Documentation

data Trie a

A trie mapping BCP47 tags to values

fromList :: [(BCP47, a)] -> Trie a

Construct a Trie from a list of tag/value pairs.

singleton :: BCP47 -> a -> Trie a

Construct a Trie from a single tag/value pair.

lookup :: BCP47 -> Trie a -> Maybe a

Lookup the most relevant item for a tag

match :: BCP47 -> Trie a -> Maybe a

Lookup an exact match for a tag

elem :: BCP47 -> Trie a -> Bool

Check if a tag exists in the Trie

union :: Trie a -> Trie a -> Trie a

A left-biased union of two Trie structures. The left value is prefered when duplicate tags are found.

unionWith :: (a -> a -> a) -> Trie a -> Trie a -> Trie a

union with a combining function.

null :: Trie a -> Bool

Check if a Trie is empty